+++ library/src/actions/readonly/readonly.ts
@@ -1,10 +1,20 @@
-import type { BaseTransformation } from '../../types/index.ts';
+import type { BaseTransformation, SuccessDataset } from '../../types/index.ts';
+
+/**
+ * Readonly output type.
+ */
+type ReadonlyOutput<TInput> =
+  TInput extends Map<infer TKey, infer TValue>
+    ? ReadonlyMap<TKey, TValue>
+    : TInput extends Set<infer TValue>
+      ? ReadonlySet<TValue>
+      : Readonly<TInput>;
 
 /**
  * Readonly action interface.
  */
 export interface ReadonlyAction<TInput>
-  extends BaseTransformation<TInput, Readonly<TInput>, never> {
+  extends BaseTransformation<TInput, ReadonlyOutput<TInput>, never> {
   /**
    * The action type.
    */
@@ -20,15 +30,17 @@ export interface ReadonlyAction<TInput>
  *
  * @returns A readonly action.
  */
+export function readonly<TInput>(): ReadonlyAction<TInput>;
+
 // @__NO_SIDE_EFFECTS__
-export function readonly<TInput>(): ReadonlyAction<TInput> {
+export function readonly(): ReadonlyAction<unknown> {
   return {
     kind: 'transformation',
     type: 'readonly',
     reference: readonly,
     async: false,
     '~run'(dataset) {
-      return dataset;
+      return dataset as SuccessDataset<Readonly<unknown>>;
     },
   };
 }